requirements.txt
tensorflow-gpu==2.0.0-beta1
imutils
matplotlib
numpy
tqdm
torch
fire
lmdb
natsort
tensorflow-serving-api
gin-config
opencv-python
shapely
drive-cli
